[kernel] E: supporting 64bit kernel cals

+++ src/kernel/src/kcall/dispatcher.rs
@@ -51,7 +51,7 @@ use ::sys::{
 /// - `number`: Number of the kernel call.
 ///
 #[no_mangle]
-pub extern "C" fn do_kcall(number: u32, arg0: u32, arg1: u32, arg2: u32, arg3: u32) -> i32 {
+pub extern "C" fn do_kcall(number: u32, arg0: u32, arg1: u32, arg2: u32, arg3: u32) -> i64 {
     let pid: ProcessIdentifier = match unsafe { ProcessManager::get() }.get_pid() {
         Ok(pid) => pid,
         Err(e) => return KcallResult::Error(e.code.into()).into(),
